- tighter visibility of classes. - refactored the way the classes where created. only package private visibility now. - change inspections settings for the project
