+++ src/org/intellij/ideaplugins/tabswitch/component/Components.java
@@ -1,0 +1,28 @@
+package org.intellij.ideaplugins.tabswitch.component;
+
+import javax.swing.JComponent;
+import javax.swing.JLabel;
+import javax.swing.JList;
+
+import com.intellij.openapi.project.Project;
+
+/**
+ * <br> User: must <br> Date: 2013-12-10
+ */
+public final class Components {
+
+  private Components() {
+  }
+
+  public static JList newList(Project project, JLabel pathLabel) {
+    return new ListComponentFactory(project).create(pathLabel);
+  }
+
+  public static JComponent newListFooter(JLabel pathLabel) {
+    return new FooterComponentFactory().create(pathLabel);
+  }
+
+  public static JLabel newPathLabel() {
+    return new PathLabelComponentFactory().create();
+  }
+}
