+++ src/org/intellij/ideaplugins/tabswitch/component/FooterComponentFactory.java
@@ -1,36 +1,22 @@
 package org.intellij.ideaplugins.tabswitch.component;
 
 import java.awt.BorderLayout;
+import java.awt.Component;
 import java.awt.Graphics;
 
 import javax.swing.BorderFactory;
 import javax.swing.JComponent;
-import javax.swing.JLabel;
 import javax.swing.JPanel;
 
 import com.intellij.ui.Gray;
 
 /**
  * Creates the footer component. It is here where the file path is displayed - this is the container class.
- * <pre>
- * User: must
- * Date: 2012-05-30
- * </pre>
  */
-public class FooterComponentFactory implements ComponentFactory<JComponent> {
+class FooterComponentFactory {
 
-  private JLabel pathLabel;
-
-  @Override
-  public JComponent create() {
-    JComponent footer = new JPanel(new BorderLayout()) {
-      @Override
-      protected void paintComponent(Graphics g) {
-        super.paintComponent(g);
-        g.setColor(Gray._135);
-        g.drawLine(0, 0, getWidth(), 0);
-      }
-    };
+  JComponent create(Component pathLabel) {
+    JComponent footer = newFooterJPanel();
 
     footer.setBorder(BorderFactory.createEmptyBorder(4, 4, 4, 4));
 
@@ -39,8 +25,14 @@ public class FooterComponentFactory implements ComponentFactory<JComponent> {
     return footer;
   }
 
-  public FooterComponentFactory withPathLabel(JLabel pathLabel) {
-    this.pathLabel = pathLabel;
-    return this;
+  private JComponent newFooterJPanel() {
+    return new JPanel(new BorderLayout()) {
+      @Override
+      protected void paintComponent(Graphics g) {
+        super.paintComponent(g);
+        g.setColor(Gray._135);
+        g.drawLine(0, 0, getWidth(), 0);
+      }
+    };
   }
 }
